+++ src/app/services/sdk/messages/chat.seal_pb.d.ts
@@ -39,6 +39,8 @@
 /* tslint:disable */
 /* tslint:disable */
 /* tslint:disable */
+/* tslint:disable */
+/* tslint:disable */
 // package: msg
 // file: chat.seal.proto