+++ src/app/services/sdk/messages/chat.seal_pb.js
@@ -52,6 +52,8 @@
 /* eslint-disable */
 /* eslint-disable */
 /* eslint-disable */
+/* eslint-disable */
+/* eslint-disable */
 // source: chat.seal.proto
 /**
  * @fileoverview