+++ public/changelog.md
@@ -1,3 +1,7 @@
+## 1.0.53
+* Call participant promote/demote implemented
+* Minor improvements in Team Panel
+
 ## 1.0.52
 * Minor improvements in update manager and left panel